# docs/swagger.yaml
basePath: /
definitions:
  domain.Interview:
    properties:
      candidate_id:
        description: Foreign key referencing the candidate's ID
        type: integer
      feedback:
        description: Feedback or notes about the interview
        type: string
      id:
        description: Unique identifier for the interview
        type: integer
      interview_date:
        description: Date and time of the interview
        type: string
      job_id:
        description: Foreign key referencing the job's ID
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API for managing interviews in the system.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Interview Service API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns the health status of the interview service
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check service health
      tags:
      - Health
  /interviews:
    get:
      description: Retrieve a list of all interviews in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of interviews
          schema:
            items:
              $ref: '#/definitions/domain.Interview'
            type: array
        "500":
          description: Failed to fetch interviews
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all interviews
      tags:
      - Interviews
    post:
      consumes:
      - application/json
      description: Add a new interview by providing candidate_id, job_id, interview_date,
        and feedback
      parameters:
      - description: Interview Creation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Interview'
      produces:
      - application/json
      responses:
        "201":
          description: Interview created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create interview
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new interview
      tags:
      - Interviews
swagger: "2.0"